Start nu-path integration tests
